feat: Add PropertiesList component to display a list of properties
This commit adds a new component called PropertiesList, which is responsible for rendering a list of properties. It imports the PropertyCard component from the card directory and uses the PropertyCardProps type from the utils/types module. The PropertiesList component takes an array of PropertyCardProps objects as a prop and renders each property using the PropertyCard component.